完善淘宝商品拉取

+++ apis/tao-api/src/main/java/cn/qihangerp/open/tao/domain/OmsTaoGoods.java
@@ -106,7 +106,7 @@ public class OmsTaoGoods implements Serializable {
     /**
      * 宝贝所属的运费模板ID，如果没有返回则说明没有使用运费模板
      */
-    private Integer postageId;
+    private Long postageId;
 
     /**
      * 商家外部编码(可与商家外部系统对接)。需要授权才能获取。
